Integrate ecocredits module (#153)
* Integrate ecocredits module

* use interfaces for Msg/Query services

* simplify decimal parsing

* use protobuf server interfaces

* Refactor the MsgServer usage with modules

I found that we are recreating a server and hardcoding it in local functions
(eg in SDK modules in all handler methods). This leads to difficulties or weird designs
if we want to test or customize it. Instead, I propose to relay more on tests and
configurability. In this commit, I compose a Server inside the AppModule. The AppModule
constructor could take a server as a parameter, however I decided to not do it and not expose
the default functionality.
Module test suites however, will like to customize it. They can do that by creating an
AppModule themselves - the struct is public.

* clean module.go

* update app

* Add Query CLI

* move client utility functions

* Adding MsgClient CLI: wip

* remove Server interface

* add AppModuleBasic to enable interface types registration

* fix tests

* grammar check

* fix imports

* fix lint

Co-authored-by: Aaron Craelius <aaronc@users.noreply.github.com>